Move the lexifier to it's own file
