+++ frontend/src/metabase-lib/v1/expressions/pratt/index.ts
@@ -1,3 +1,4 @@
+export * from "./lexifier";
 export * from "./parser";
 export * from "./compiler";
 export * from "./syntax";
